Track - TrackReference utils moved to class Tracks
